// src/types/section.ts
export type SectionKey =
	| 'Home'
	| 'MyClub'
	| 'ClubMembers'
	| 'FighterInfo'
	| 'CombatList'
	| 'Championship'
	| 'Clubs'
	| 'PlayerClubDetails'
	| 'ClubThings'
	| 'Base'
	| 'FriendlyCombats'
	| 'Market'
	| 'Masters'
	| 'Repair'
	| 'ClubReview'
	| 'Donate'
	| 'FreeFighters'
	| 'Settings'
	| 'Transfer'
	| 'Chat';

export type FrameDivTarget = 'Left' | 'Center' | 'Right';